remove prints and add missing fields

+++ types/protobuf.go
@@ -85,7 +85,7 @@ func (tm2pb) BlockID(blockID BlockID) tmproto.BlockID {
 
 func (tm2pb) PartSetHeader(header PartSetHeader) tmproto.PartSetHeader {
 	return tmproto.PartSetHeader{
-		Total: uint32(header.Total),
+		Total: header.Total,
 		Hash:  header.Hash,
 	}
 }
